Refactor: Create SchemaReader for schema migration related functions
